Publish as ESM-only (#11)
Require Node v20.19 (which enables require(esm) support by default).

This new package is supposed to be compatible with Apollo Server 4 which
does an elaborate dance to publish as both CJS and ESM, but we already
decided it was OK for this package itself to require Node v20.

Turns out CJS-only plays poorly with the AS4-style dual build:
https://github.com/apollographql/apollo-server/issues/8069

And we're likely to go ESM-only with AS5 because newer Node makes this
easier. So let's do that here too.

Making tests work requires:
- Tweaking ts-jest to some new presets.
- Passing an experimental option to Node (and disabling the warning
  about it).
- Upgrading the integration test suite to a version that is dual-built
  instead of CJS-only (otherwise we end up with different ApolloServers
  from the `@apollo/server` we load directly (ESM) and via the
  integration test suite (CJS)).